+++ src/components/TodoList.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { addTodo } from '../actions/todoActionCreator';
-import Form from './Form/Form';
 
 const TodoList = (props) => {
     const todoItems = props.todos.map( (todo, index) => {
@@ -13,7 +11,6 @@ const TodoList = (props) => {
     });
     return (
         <div>
-            <Form onClickHandler={ (text) => {props.addTodo(text);} }/>
             {todoItems}
         </div>
     );
@@ -23,16 +20,6 @@ const mapStateToProps = (state) => {
     return { todos: state.todos };
 };
 
-const mapDispatchToProps = (dispatch) => {
-    return {
-        addTodo: (text) => {
-            const addAction = addTodo(text);
-            dispatch(addAction);
-        }
-    };
-};
-
 export default connect(
-    mapStateToProps,
-    mapDispatchToProps
+    mapStateToProps
 )(TodoList);
